spec-368: neutral shared-formatting module + split assessPhaseTransition [P3]
Behaviour-preserving refactor combining audit (spec-345) findings sol-4 + sol-5
under umbrella spec-354. Output is byte-identical.

sol-4 (std-12): promote the shared formatting + phase-guidance helpers out of
mcp/formatters.ts into the neutral module formatting/formatters.ts. mcp/formatters.ts
is now a thin barrel re-export so mcp/-internal call sites are unchanged;
cross-boundary importers (agent/handlers/*, agent/context-builder, services/*) now
depend on the neutral path. phaseFromStatus is defined once and exported from the
neutral module; context-builder's duplicate switch is deleted (statusToPhase
delegates, keeping its 'specify' fallback).

sol-5: move formatPhaseAssessment (presentation) into the neutral module
(formatting/phase-assessment-format.ts), re-exported from services/phase-assessment.ts;
split assessPhaseTransition into per-rule nudge predicate functions composed by a
thin orchestrator.

Source-scanning regression guards (footer-one-seat, spec-193-tripwire, spec-219-*,
b105-ac-coverage, decisions-need-acs, nudge-parity, ac-nag-sketch, formatters.test)
repointed to the new module paths (assertions unchanged); legacy-spec-vocab
allowlist entry repointed.

tsc: only the ~14 pre-existing unrelated test-file errors. make test-server:
4301 passed; only the env-gated spec-129-ci-emission-key and the known
migration-smoke flake fail. Biome unused-import sweep over touched files: zero.
