+++ packages/server/src/__regression__/spec-219-lockstep-203.regression.test.ts
@@ -29,7 +29,8 @@ const toolSpecs = [
     .filter((n) => n.endsWith(".ts"))
     .map((n) => readFileSync(join(HANDLERS_DIR, n), "utf-8")),
 ].join("\n");
-const formatters = read(join("mcp", "formatters.ts"));
+// spec-368 sol-4: formatter body moved to the neutral module; scan it there.
+const formatters = read(join("formatting", "formatters.ts"));
 
 describe("spec-219 ac-13 — the per-tool seat inverts spec-203 ac-13", () => {
   it("the seat composes the envelope PER-TOOL (tool-name conditionals)", () => {
