+++ packages/server/src/mcp/ac-nag-sketch.test.ts
@@ -24,7 +24,13 @@ const SERVER_SRC = resolve(__dirname, "..");
 const SPEC = "mindset-prod/memex-building-itself/specs/spec-121";
 const acRef = (n: number) => `${SPEC}/acs/ac-${n}`;
 
-const formattersSrc = readFileSync(resolve(__dirname, "formatters.ts"), "utf8");
+// spec-368 sol-4: formatter body moved to the neutral module; mcp/formatters.ts
+// is now a thin barrel. Scan the new home so these source assertions still see
+// the code (assertions unchanged).
+const formattersSrc = readFileSync(
+  resolve(__dirname, "..", "formatting", "formatters.ts"),
+  "utf8",
+);
 const sketchSrc = readFileSync(resolve(__dirname, "ac-test-sketch.ts"), "utf8");
 // spec-366: the create_ac handler + the resolve_decision sketch wiring moved
 // from tool-specs.ts into agent/handlers/*.ts (shared infra in shared.ts). Scan
